+++ lib/ag_region_monitor_method_channel.dart
@@ -145,6 +145,26 @@ class MethodChannelAgRegionMonitor extends AgRegionMonitorPlatform {
   }
 
   @override
+  Future<void> setNotificationsRepeatEnabled(bool enabled) async {
+    try {
+      await _channel.invokeMethod('setNotificationsRepeatEnabled', {'enabled': enabled});
+    } catch (e) {
+      debugPrint('Error setting notification state: $e');
+      rethrow;
+    }
+  }
+
+  @override
+  Future<void> setNotificationsRepeatTimer(int timer) async {
+    try {
+      await _channel.invokeMethod('setNotificationsRepeatTimer', {'timer': timer});
+    } catch (e) {
+      debugPrint('Error setting notification state: $e');
+      rethrow;
+    }
+  }
+
+  @override
   Stream<Map<String, dynamic>> get regionEvents {
     return _regionEventChannel.receiveBroadcastStream().map((event) {
       if (event is Map) {