+++ CHANGELOG.md
@@ -10,3 +10,9 @@
 
 * Local notifications title and body can be customized for each region.
 * Enable or Dislable local notifications.
+
+## 1.0.4
+
+* Added support for repeating notifications when the user enters a region
+* Added ability to enable or disable repeating notifications
+* Added option to configure repeat interval for notifications